abstract `fetchAccounts` to context
